For now ensure Input.__iter__ returns CodeChunk
The GitDiff input does not necessarily return whole files and needs to set
the start and end lines.

Ideally we would be able to represent whole files in the prompt seperately from
CodeChunks, but this needs to be done a different way.

For now, this isn't strictly necessary, so reverting it to the old behavior.