workflows/code: update to reflect changes in core (#105)
* workflows/code: use filesystem tools instead of treesitter tools

The models have had trouble calling the tree-sitter-based tools
correctly. Switch to the filesystem tools, which are similar to those
used by Claude Code and Cursor, since the models can use those
reliably.

* workflow/code: move step initialization into the constructor

Previously, the CLI args were passed to the workflow `run` method, not
its constructor. Steps that depended on those args had to be
initialized lazily, not in the constructor. Now all args are available
to the constructor, so we can construct all the steps there.