litellm: fix crash on final iteration
We limit the number of iterations (turns) in an LLM conversation by
attempting to require a final answer, not another tool call, after a
set number of iterations. Unfortunately, preventing tool calls on the
final iteration across all models/providers is not that
straightforward.

The existing apporach was to exclude the `tools` param (the list tool
schemas) from the completion params. This was unreliable with Claude
(Sonnet 4.5); it would still sometimes return another tool call.

Setting the `tools` param to an explicit empty list `[]` and settings
`tool_choice` to `NONE` fixed the issue with Claude (and still worked
with OpenAI GPT-5). However, it broke Gemini (2.5 Pro).

So our final approach now is to:

1. Pass an explicit empty tools schema `tools = []` ("There are no tools to call.").
2. Pass an explicit `tool_choice = None` ("Do not call any tools".)
3. Give the LLM an extra three attempts (iterations) to issue a final response.

So far, in my testing, a single extra attempt has been
sufficient. However, if the LLM were to fail to return a final
response even with all three extra attempts, raise an exception as we
did before.